Separate out testing functions into separate libraries. Move most things to internal.
