+++ extractor/filesystem/language/python/pipfilelock/extractor_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/google/osv-scalibr/extractor"
 	"github.com/google/osv-scalibr/extractor/filesystem/language/python/pipfilelock"
 	"github.com/google/osv-scalibr/extractor/filesystem/osv"
-	"github.com/google/osv-scalibr/testing/sharedtesthelpers"
+	"github.com/google/osv-scalibr/testing/extracttest"
 )
 
 func TestExtractor_FileRequired(t *testing.T) {
@@ -64,24 +64,24 @@ func TestExtractor_FileRequired(t *testing.T) {
 func TestExtractor_Extract(t *testing.T) {
 	t.Parallel()
 
-	tests := []sharedtesthelpers.TestTableEntry{
+	tests := []extracttest.TestTableEntry{
 		{
 			Name: "invalid json",
-			InputConfig: sharedtesthelpers.ScanInputMockConfig{
+			InputConfig: extracttest.ScanInputMockConfig{
 				Path: "testdata/not-json.txt",
 			},
 			WantErrContaining: "could not extract from",
 		},
 		{
 			Name: "no packages",
-			InputConfig: sharedtesthelpers.ScanInputMockConfig{
+			InputConfig: extracttest.ScanInputMockConfig{
 				Path: "testdata/empty.json",
 			},
 			WantInventory: []*extractor.Inventory{},
 		},
 		{
 			Name: "one package",
-			InputConfig: sharedtesthelpers.ScanInputMockConfig{
+			InputConfig: extracttest.ScanInputMockConfig{
 				Path: "testdata/one-package.json",
 			},
 			WantInventory: []*extractor.Inventory{
@@ -97,7 +97,7 @@ func TestExtractor_Extract(t *testing.T) {
 		},
 		{
 			Name: "one package dev",
-			InputConfig: sharedtesthelpers.ScanInputMockConfig{
+			InputConfig: extracttest.ScanInputMockConfig{
 				Path: "testdata/one-package-dev.json",
 			},
 			WantInventory: []*extractor.Inventory{
@@ -113,7 +113,7 @@ func TestExtractor_Extract(t *testing.T) {
 		},
 		{
 			Name: "two packages",
-			InputConfig: sharedtesthelpers.ScanInputMockConfig{
+			InputConfig: extracttest.ScanInputMockConfig{
 				Path: "testdata/two-packages.json",
 			},
 			WantInventory: []*extractor.Inventory{
@@ -137,7 +137,7 @@ func TestExtractor_Extract(t *testing.T) {
 		},
 		{
 			Name: "two packages alt",
-			InputConfig: sharedtesthelpers.ScanInputMockConfig{
+			InputConfig: extracttest.ScanInputMockConfig{
 				Path: "testdata/two-packages-alt.json",
 			},
 			WantInventory: []*extractor.Inventory{
@@ -161,7 +161,7 @@ func TestExtractor_Extract(t *testing.T) {
 		},
 		{
 			Name: "multiple packages",
-			InputConfig: sharedtesthelpers.ScanInputMockConfig{
+			InputConfig: extracttest.ScanInputMockConfig{
 				Path: "testdata/multiple-packages.json",
 			},
 			WantInventory: []*extractor.Inventory{
@@ -201,7 +201,7 @@ func TestExtractor_Extract(t *testing.T) {
 		},
 		{
 			Name: "package without version",
-			InputConfig: sharedtesthelpers.ScanInputMockConfig{
+			InputConfig: extracttest.ScanInputMockConfig{
 				Path: "testdata/no-version.json",
 			},
 			WantInventory: []*extractor.Inventory{},
@@ -213,7 +213,7 @@ func TestExtractor_Extract(t *testing.T) {
 		t.Run(tt.Name, func(t *testing.T) {
 			t.Parallel()
 			e := pipfilelock.Extractor{}
-			_, _ = sharedtesthelpers.ExtractionTester(t, e, tt)
+			_, _ = extracttest.ExtractionTester(t, e, tt)
 		})
 	}
 }
