+++ testing/internal/errormatcher/errormatcher.go
@@ -1,0 +1,35 @@
+package errormatcher
+
+import (
+	"errors"
+	"strings"
+	"testing"
+)
+
+// ExpectErrContaining checks if a error contains a certain string, if not fail the test
+func ExpectErrContaining(t *testing.T, err error, str string) {
+	t.Helper()
+
+	if err == nil {
+		t.Errorf("Expected to get error, but did not")
+		return
+	}
+
+	if !strings.Contains(err.Error(), str) {
+		t.Errorf("Expected to get \"%s\" error, but got \"%v\"", str, err)
+	}
+}
+
+// ExpectErrIs checks if a error is another error, if not fail the test
+func ExpectErrIs(t *testing.T, err error, expected error) {
+	t.Helper()
+
+	if err == nil {
+		t.Errorf("Expected to get error, but did not")
+		return
+	}
+
+	if !errors.Is(err, expected) {
+		t.Errorf("Expected to get \"%v\" error but got \"%v\" instead", expected, err)
+	}
+}
